+++ keyfactor/data_source_keyfactor_certificate.go
@@ -283,7 +283,7 @@ func (r dataSourceCertificate) Read(ctx context.Context, request tfsdk.ReadDataS
 			password = generatePassword(32, 4, 4, 4)
 		}
 		tflog.Info(ctx, "Requested certificate has a private key attempting to recover from Keyfactor Command.")
-		pKeyO, _, chainO, dErrO := r.p.client.RecoverCertificate(cResp.Id, "", "", "", password)
+		pKeyO, _, chainO, dErrO := r.p.client.RecoverCertificate(cResp.Id, "", "", "", password, collectionIdInt)
 		if dErrO != nil {
 			tflog.Error(ctx, fmt.Sprintf("Unable to recover private key for certificate '%v' from Keyfactor Command.", cResp.Id))
 			response.Diagnostics.AddWarning(